requirements.txt
streamlit
pandas
geopandas
folium
streamlit-folium
scikit-learn
xgboost
openai
langchain
faiss-cpu
python-dotenv
fastapi 
uvicorn 
psycopg2-binary
pgvector
sentence-transformers
langchain
python-dotenv
pydantic
schedule
selenium
webdriver-manager
pydantic-settings
langchain-groq
matplotlib
numpy
requests
